only remove files that exist

+++ cctbx/update_bootstrap.sh
@@ -6,6 +6,9 @@ alias this="readlink -f \$(dirname \${BASH_SOURCE[0]})"
 
 
 pushd $(this)
-rm bootstrap.py
+if [[ -f bootstrap.py ]]
+then
+    rm bootstrap.py
+fi
 wget "https://raw.githubusercontent.com/cctbx/cctbx_project/master/libtbx/auto_build/bootstrap.py" --no-check-certificate
 popd
